+++ sdk/trace/batch_span_processor_test.go
@@ -36,11 +36,22 @@ type testBatchExporter struct {
 	sizes         []int
 	batchCount    int
 	shutdownCount int
+	delay         time.Duration
+	err           error
 }
 
 func (t *testBatchExporter) ExportSpans(ctx context.Context, ss []*export.SpanSnapshot) error {
 	t.mu.Lock()
 	defer t.mu.Unlock()
+
+	time.Sleep(t.delay)
+
+	select {
+	case <-ctx.Done():
+		t.err = ctx.Err()
+		return ctx.Err()
+	default:
+	}
 
 	t.spans = append(t.spans, ss...)
 	t.sizes = append(t.sizes, len(ss))
@@ -88,22 +99,34 @@ func TestNewBatchSpanProcessorWithNilExporter(t *testing.T) {
 }
 
 type testOption struct {
-	name           string
-	o              []sdktrace.BatchSpanProcessorOption
-	wantNumSpans   int
-	wantBatchCount int
-	genNumSpans    int
-	parallel       bool
+	name              string
+	o                 []sdktrace.BatchSpanProcessorOption
+	wantNumSpans      int
+	wantBatchCount    int
+	wantExportTimeout bool
+	genNumSpans       int
+	delayExportBy     time.Duration
+	parallel          bool
 }
 
 func TestNewBatchSpanProcessorWithOptions(t *testing.T) {
 	schDelay := 200 * time.Millisecond
+	exportTimeout := time.Millisecond
 	options := []testOption{
 		{
 			name:           "default BatchSpanProcessorOptions",
 			wantNumSpans:   2053,
 			wantBatchCount: 4,
 			genNumSpans:    2053,
+		},
+		{
+			name: "non-default ExportTimeout",
+			o: []sdktrace.BatchSpanProcessorOption{
+				sdktrace.WithExportTimeout(exportTimeout),
+			},
+			wantExportTimeout: true,
+			genNumSpans:       2053,
+			delayExportBy:     2 * exportTimeout, // to ensure export timeout
 		},
 		{
 			name: "non-default BatchTimeout",
@@ -171,7 +194,9 @@ func TestNewBatchSpanProcessorWithOptions(t *testing.T) {
 	}
 	for _, option := range options {
 		t.Run(option.name, func(t *testing.T) {
-			te := testBatchExporter{}
+			te := testBatchExporter{
+				delay: option.delayExportBy,
+			}
 			tp := basicTracerProvider(t)
 			ssp := createAndRegisterBatchSP(option, &te)
 			if ssp == nil {
@@ -185,16 +210,21 @@ func TestNewBatchSpanProcessorWithOptions(t *testing.T) {
 			tp.UnregisterSpanProcessor(ssp)
 
 			gotNumOfSpans := te.len()
-			if option.wantNumSpans != gotNumOfSpans {
+			if option.wantNumSpans > 0 && option.wantNumSpans != gotNumOfSpans {
 				t.Errorf("number of exported span: got %+v, want %+v\n",
 					gotNumOfSpans, option.wantNumSpans)
 			}
 
 			gotBatchCount := te.getBatchCount()
-			if gotBatchCount < option.wantBatchCount {
+			if option.wantBatchCount > 0 && gotBatchCount < option.wantBatchCount {
 				t.Errorf("number batches: got %+v, want >= %+v\n",
 					gotBatchCount, option.wantBatchCount)
 				t.Errorf("Batches %v\n", te.sizes)
+			}
+
+			if option.wantExportTimeout && te.err != context.DeadlineExceeded {
+				t.Errorf("context deadline: got err %+v, want %+v\n",
+					te.err, context.DeadlineExceeded)
 			}
 		})
 	}